Minor refactoring: moved SegmentVisitor to Visitors directory
